Wrap initialization of bools' values in try-catch
Android versions before Nougat do not have `getByte` method in `sun.misc.Unsafe`. This causes loading of LongHashFunction to trigger NoSuchMethodError when running on older versions of platform (see #18 for details).

The method in question is currently being used to obtain *raw integer value of boolean*. The proposed workaround adds a fallback to hardcoded values of 1 and 0. While this might not be the most reliable way to do it, this fixes the crash on older Android versions (and I personally don't use Zero-Allocation-Hashing with boolean arrays anyway).

I don't think that there is a better way to fix this. Current approach to hashing boolean values is already imperfect and makes a number of shaky implicit assumptions: it assumes that size of boolean is exactly one byte (was not always true in the past), assumes the specific layout of boolean arrays and assumes that it matches layout of individual booleans in memory. Unlike the rest of primitives, booleans do not have a corresponding `Buffer` type, and they layout in memory is left implementation dependent. Personally I believe, that the `hashBoolean()` and `hashBooleans()` should be deprecated altogether with an appropriate comment — not with intention of removing them in future, but to let people know, that quickly hashing boolean arrays (and expecting the result to  be predictable and consistent in future JVM versions) is a bit of gamble.